+++ brain/python/compartmentReport.cpp
@@ -63,6 +63,7 @@ public:
         return view->getMapping().getNumCompartments(index);
     }
 
+    size_t getFrameSize() const { return view->getMapping().getFrameSize(); }
     CompartmentReportViewPtr view;
 };
 
@@ -95,8 +96,6 @@ bp::object CompartmentReport_getMetaData(const CompartmentReport& reader)
     dict["time_step"] = md.timeStep;
     dict["time_unit"] = md.timeUnit;
     dict["data_unit"] = md.dataUnit;
-    dict["cell_count"] = md.cellCount;
-    dict["compartment_count"] = md.compartmentCount;
     dict["frame_count"] = md.frameCount;
 
     return dict;
@@ -188,6 +187,8 @@ bp::class_<CompartmentReport, boost::noncopyable>(
                   DOXY_FN(brain::CompartmentReport::getMetaData))
     .add_property("gids", CompartmentReport_getGids,
                   DOXY_FN(brain::CompartmentReport::getGIDs))
+    .add_property("cell_count", &CompartmentReport::getCellCount,
+                  DOXY_FN(brain::CompartmentReport::getCellCount))
     .def("create_view", CompartmentReport_createView,
          (selfarg, bp::arg("gids")),
          DOXY_FN(brain::CompartmentReport::createView(const GIDSet&)))
@@ -202,6 +203,8 @@ bp::class_<CompartmentReportMappingProxy>("CompartmentReportMapping",
                   DOXY_FN(brain::CompartmentReportMapping::getIndex))
     .add_property("offsets", CompartmentReportMapping_getOffsets,
                   DOXY_FN(brain::CompartmentReportMapping::getOffsets))
+    .add_property("frame_size", &CompartmentReportMappingProxy::getFrameSize,
+                  DOXY_FN(brain::CompartmentReportMapping::getFrameSize))
     .def("compartment_counts",
          CompartmentReportMapping_getCompartmentCounts, (selfarg),
          DOXY_FN(brain::CompartmentReportMapping::getCompartmentCounts));